+++ CHANGES.md
@@ -7,6 +7,7 @@
 -   Testing: completed GMW-filter tests
 -   Testing: extent GMN-models tests
 -   Testing: extent GMN-serializers tests
+-   Testing: completed api-utils tests
 
 
 ## 1.59 (2025-12-01)